requirements.txt
flask
flask-sqlalchemy
psycopg2-binary
werkzeug
python-dotenv
flask-session
flask-login
flask-cors